==> Notebooks/2b_ETL_UserItems.ipynb <==
# 1b - ETL Archivo User_Items

# Librerias a Usar
import pandas as pd
import ast
import fun

# abrir el csv en un DataFrame
df_items=pd.read_csv(r'./user_items.csv ')

df_items.head(2)

- Se explicara a groso modo cada columna : &nbsp;

    - **item_id&emsp;:**&emsp;Identificador único para el juego.
    - **item_name&emsp&emsp;:**&emsp;Nombre del juego.
    - **playtime_2weeks&emsp;:**&emsp;Tiempo jugado en las ultimas 2 semanas.
    - **steam_id&emsp;:**&emsp;Identificador único para jugador.
    - **items_count&emsp;:**&emsp;Cantidad de items en el juego por usuarios.
    - **user_id&emsp;:**&emsp;Identificador único para jugador.
    - **user_url&emsp;:**&emsp;Direccion del sitio web del juego.
   
</br>
</br>
- Las columnas 'url', 'reviews_url' son columnas innecsarias por ende se eliminaran


# Se verifica si existen nulos y el tipo de data de cada columna
fun.tipo_dato(df_items)

#### Se decide eliminar la columna  'playtime_2weeks' dado que existe otra columna que se llama 'playtime_forever', 'items count' que es el conteo de items por usuario que no es necesaria y la de 'user_url' que tampoco se observa nevesaria

# Se eliminan algunas columnas innecesarias
df_items.drop(columns=['playtime_2weeks','items_count','user_url'],inplace=True)

# Se verifican todas las filas que contengan nulos 
nulos = df_items[df_items.isnull().any(axis=1)]
nulos

# Se revisan si existen filas duplicadas en el DataFrame 
duplicados = df_items.loc[df_items.duplicated()]
duplicados

# se eliminan los 59117 datos duplicados
df_items = df_items.drop_duplicates(keep='first') # se conserva la primera aparición de esa fila y se eliminan las apariciones duplicadas subsiguientes.

df_items

# Cambiar nombre a la columna 'item_id' 
df_items.rename(columns={'item_id' : 'id_game'}, inplace=True)

# Ordenamos los id's en forma ascendente y hacemos un reset al index 
df_items=df_items.sort_values(by='id_game', ascending=True)
df_items.reset_index(drop=True, inplace=True)

df_items

# # Capitalizar la primera letra de cada palabra en el nombre de las columnas
def capitalizar(texto):
    return texto.capitalize()
df_items.columns = df_items.columns.map(capitalizar)

# Exportamos a csv
df_items.to_csv('items.csv', index=False)
