==> Notebooks/2c_ETL_UserReviews.ipynb <==
import pandas as pd
import ast
import fun
import re

# abrir el csv en un DataFrame
df_reviews=pd.read_csv(r'./user_reviews.csv ')

df_reviews.head(3)

- Dado que se debe crear una columna  'sentiment_analysis' usando las reseñas de juegos hechas por los usuarios es nesario depurar algo de informacion de esta columna 'reviews ', por ende se explicara a groso modo cada columna : 

    - **user_id&emsp;:**&emsp;Identificador único para el usuario.
    - **user_url&emsp;:**&emsp;Direccion url del perfil del usuario en streamcommunity.
    - **item_id&emsp;:**&emsp;Identificador único del item que seria el id del juego.
    - **funny&emsp;:**&emsp;Número de personas que marcaron esto comenta que es gracioso.
    - **posted&emsp;:**&emsp;Fecha de posteo del review.
    - **last_edited&emsp;:**&emsp;Fecha de la última edición del review.
    - **helpful:&emsp;:**&emsp;Número de personas que marcaron el review ('reseña') como útil.
    - **recommend&emsp;:**&emsp;Recomendado por el usuario (Si o No = True o False).
    - **review&emsp;:**&emsp;Texto con la reseña sobre un juego escrita por un usuario. 
</br>
</br>
- Las columnas 'reviews_funny', 'reviews_last_edited' y 'helpful' son columnas innecesarias puesto que 'funny' aplica solo si el comentario fue grasioso para otros usuarios, 'last_edited' es la fecha de la ultima edicion de la reseña y 'helpful' valora si la reseña de un usuario fue de ayuda para otros usuarios pero existen muchas filas que tienen valores como 'No ratings yet', es decir sin valoracion por cual no es necesria dicha columna.
- La columna 'user_url' es innecesaria, se eliminara

- En la columna 'posted' ahi que quitarle la palabra 'Posted' que aparece antes de la fecha, de igual manera las fechas es mejor manejalos con  formato YYYY-MM-DD

# Eliminacion de columnas
df_reviews = df_reviews.drop(columns=['funny','last_edited','helpful','user_url'])
df_reviews.columns

# Se verifica si existen valores nulos y el tipo de dato de cada columna
fun.tipo_dato(df_reviews)

# Se revisan si existen filas duplicadas en el DataFrame 
duplicados =  df_reviews.loc[df_reviews.duplicated()]
duplicados

# se eliminan los 874 datos duplicados
df_reviews1 = df_reviews.drop_duplicates(keep='first') # se conserva la primera aparición de esa fila y se eliminan las apariciones duplicadas subsiguientes.

### Columna 'item_id'

# Se renombra la columna 'item_id'
df_reviews = df_reviews.rename(columns={'item_id': 'id_game'})
df_reviews['id_game'] = df_reviews['id_game'].astype(int)
df_reviews.columns

### Columna 'posted'

df_reviews['posted'].unique()

Se observa la palabra 'posted' al inicio de cada fecha, se procede a cambiar la fecha a un formato 'yyyy-mm-dd', de igual manera solo se necesita solamente el año y existen datos que no tienen por ende se eliminaran dado que no se podria generar una fecha exacta para los susuarios que genraron dicho post entonces se opta por la eliminacion

# Funcion para convertir fecha

def convertir_fecha(cadena_fecha):
#\w+: es una clase de caracteres que representa caracteres alfabéticos (letras, números o guiones bajos), y el + significa una o más ocurrencias.
#\s: Esto coincide con un espacio en blanco.
#\d{1,2}: \d representa cualquier dígito (0-9), y {1,2} significa que esperamos uno o dos dígitos.

    match = re.search(r'(\w+\s\d{1,2},\s\d{4})', cadena_fecha) # Busca un patrón en la cadena de fecha que coincida con el formato "Mes Día, Año".
    if match:
        fecha_str = match.group(1) # Extrae el texto coincidente, que es la fecha en formato "Mes Día, Año".
        try:
            fecha_dt = pd.to_datetime(fecha_str) # Intenta convertir la fecha en un objeto datetime de Pandas.
            return fecha_dt.strftime('%Y-%m-%d')# Formatea la fecha en el formato 'YYYY-MM-DD' y devuelve el resultado.
        except:
            return 'Fecha inválida' # Si hay un error al convertir la fecha, se devuelve 'Fecha inválida'.
    else:
        return 'Formato inválido'# Si no se encuentra un patrón coincidente en la cadena, se devuelve 'Formato inválido'.

# Se convierte la fecha al formato deseado y se crea en una columna llamada 'review_date'
df_reviews['posted_date'] =df_reviews['posted'].apply(convertir_fecha) 
df_reviews.head(2)

# Se elimina la columna 'posted'
df_reviews = df_reviews.drop(columns=['posted'])
df_reviews.columns

# Cilco for para poder ver todos los valores de la columna en un texto
especif = df_reviews['posted_date'].unique()
for i in especif:
    print(i)

no_validos = df_reviews[(df_reviews['posted_date']=='Fecha inválida')|(df_reviews['posted_date']=='Formato inválido')]
no_validos

# Se eliminan los valores qque tengan 'Formato invalido'
df_reviews = df_reviews[df_reviews['posted_date'].str.strip() != 'Formato inválido']

### Columna 'recommend'

# Pasar a valores booleanos la columna 'recommend'
df_reviews.recommend = df_reviews.recommend.astype(int)
df_reviews.head(4)

### Columna 'review	'

# se pasan a str todos los valores de columna review
df_reviews['review'] = df_reviews['review'].astype(str)

# se ordenan las columnas a una manera mas legible
orden=['id_game','review','posted_date','recommend',]
df_reviews = df_reviews[orden]
df_reviews.columns

# Ordenamos los id's game en forma ascendente y hacemos un reset al index 
df_reviews=df_reviews.sort_values(by='id_game', ascending=True)
df_reviews.reset_index(drop=True, inplace=True)
df_reviews

# Se verifica que no se tenga valores vacios en la columna reviews
if df_reviews['review'].isnull().any():
    print("Hay valores nulos en la columna 'review'.")
else:
    print("No hay valores nulos en la columna 'review'.")

# Capitalizar la primera letra de cada palabra en el nombre de las columnas
def capitalizar(texto):
    return texto.capitalize()
df_reviews.columns = df_reviews.columns.map(capitalizar)
df_reviews.head(5)

df_reviews.to_csv('reviews.csv', index=False)

## Creacion de la columna 'sentiment_analysis'

Para el análisis de sentimiento se utilizara TextBlob, se calculara la polaridad de sentimiento y se clasificara de la siguiente manera  positivo, negativo o neutral, Textblob tiene un rango de -1 a 1 en sentimiento de polaridad, donde valores cercanos a 1 indican un sentimiento positivo y valores cercanos a -1 indican un sentimiento negativo; para este ejercicio las polaridades tendran un rago entre -0.2 y 0.2,es decir polaridades por debajao de -0.2 seran(negativas), por encima de 0.2 seran (positivas), y las esten en medio del rango seran (neutrales).

import nltk # Biblioteca Natural Language Toolkit
from nltk.corpus import stopwords #  Lista de palabras comunes en varios idiomas que se consideran "palabras vacías" 
from textblob import TextBlob # Biblioteca de procesamiento de lenguaje natural (NLP) e
from textblob import Word # Operaciones de procesamiento de texto en palabras individuales.
nltk.download('wordnet') # Base de datos léxica y un recurso muy útil para el procesamiento de lenguaje natural (NLP).
nltk.download('stopwords') # Conjunto de palabras vacías en el idioma inglés.

df_reviews1=df_reviews

df_reviews1['Review'] = df_reviews1['Review'].astype(str)

# se paasa a minúsculas todas las palabras de la columna
df_reviews1['Review'] = df_reviews1['Review'].str.lower()

df_reviews1

# Utiliza una expresión regular para eliminar caracteres especiales y mantener solo letras y espacio, con esto tambien se eliminaran las filas
# que estan en otro idioma y solo dejar las de habla inglesa
df_reviews1['Review'] = df_reviews1['Review'].apply(lambda x: re.sub(r'[^a-zA-Z\s]', '', x))

# Se buscan las filas vacias para luego eliminarlas
vacias = df_reviews1[~df_reviews1['Review'].str.contains(r'[a-zA-Z]')]
vacias

# Se eliminan las filas que estan vacias en la columna 'Review'
df_reviews1= df_reviews1[df_reviews1['Review'].str.contains(r'[a-zA-Z]')].reset_index()

Se pasa a eliminar las palabras stopworks del idioma ingles ya que estas son palabras comunes que no aportan mucha informacion para ello se descargaron los recursos necesarios paraesto como lo es ('wordnet') que contiene una base de datos lexica de lenguaje inglesa para la lematizacion (por ejemplo "jugando" se lematiza a  "jugar"), la desambiguación de palabras, búsqueda semántica y generación de lenguaje natural, tambien se descarga ('stopwords') el cual tiene una lista de palabras comunes de la lengua inglesa (ejemplo de stopwords: "a, an, of, for or, and, that...etc")

# Función para normalizar el texto y eliminar stopwords
def preprocess_text(text, language='english'):
    words = TextBlob(text).words
    words = [Word(word).lemmatize() for word in words if word.lower() not in stopwords.words(language)]
    return ' '.join(words)

# Se crea una nueva columna 'Normalized_review' donde contendra los valores procesados de la columna 'Review'
df_reviews1['Normalized_review'] = df_reviews1['Review'].apply(preprocess_text)

df_reviews1.head(7)

# Se crea la función de analisis de sentimiento
def sentiment_analysis(text):
    if isinstance(text, str) and len(text) > 0: # se valida que no este vacio el argumento y que sea una cadena de texto 
        analysis = TextBlob(text) # se crea una instancia de TextBlob
        sentiment_score = analysis.sentiment.polarity # 'sentiment_score' se usa para almacenar la puntuación de polaridad
        if sentiment_score < -0.2:
            return 0  # Malo
        elif sentiment_score > 0.2:
            return 2  # Positivo
        else:
            return 1  # Neutral
    else:
        return 1  # Valor por defecto si no hay reseña

# Se aplica la funcion anilisis de sentimiento a la columna 'Normalized_review'
df_reviews1['Sentiment_analysis']=df_reviews1['Normalized_review'].apply(sentiment_analysis)

# Se elimina las columnas que no se van a necesitar
df_reviews1 = df_reviews1.drop(columns=['Review','Normalized_review','index','Posted_date'])
df_reviews1.columns

df_reviews1

# Exportamos a csv
df_reviews1.to_csv('review_sentiment.csv', index=False)
