==> Notebooks/4_Ajustes_API.ipynb <==
# Preparación de conjuntos de datos para consultas

### Se preparan los csv necesarios para hacer las consultas en la API.

import pandas as pd
import numpy as np
import ast

## --> PlayTimeGenre
### ** DataFrames y Columnas  **


- A continuación se lista lo necesario para el primer endpoint :

    1. def PlayTimeGenre( genero : str ): Debe devolver año con mas horas jugadas para dicho género.
        
        - df_games.Genre = (str)
        - df_games.Release_year = (int)
        - df_games.Id_game = (int)
        - df_items.Playtime_forever = (str)
             
        A partir de los dataset obtener un nuevo conjunto de datos unicamente con los siguientes campos que necesitará la consulta:
        - genre (str) = genero unico dentro de la lista total de generos
        - playtime_hours (int) = cantidad de horas de juego totales
        - year (int) = año de lanzamiento del juego 


# Se cargan los DataFrame necesarios
df_games = pd.read_csv("./data/games.csv")
df_items = pd.read_csv("./data/items.csv")


df_games.head(2)

df_items.head(2)

# Se crean los dataframes a utilizar para luego hacer el dataframe final para la consulta
df_1 = df_games[['Id_game','Genres','Release_year']].copy()
df_2 = df_items[['Id_game','Playtime_forever']].copy()

# Se realiza un inner join 
df_genre =  df_1.merge(df_2, on='Id_game', how='inner')
df_genre.head(2)

# Se elimina la columna 'Id_game' ya que no es necesaria
df_genre = df_genre.drop(columns='Id_game')
df_genre.columns

# Se cambian los valores 'No_Data' por 0
df_genre['Release_year'] = df_genre['Release_year'].replace('No_Data', 0).astype('int16')

# se eliminan los valores con 0
df_genre = df_genre[df_genre['Release_year'] != 0]

# Se eliminar registros en la columna 'playtime_forever' donde el valor es igual a 0
df_genre = df_genre[df_genre["Playtime_forever"] != 0]
df_genre = df_genre.reset_index(drop=True)

df_genre.head(2)

# Convertir las cadenas tipo lista en listas reales
df_genre['Genres'] = df_genre['Genres'].apply(ast.literal_eval)

# se desglosa la columna 'Genre' en filas separadas 
df_genre = df_genre.explode('Genres')

# Se suma el conteo de 'playtime_forever' agrupando por 'Genre' y 'Release_year' 
df_final_genre = df_genre.groupby(['Genres', 'Release_year'])['Playtime_forever'].sum().reset_index()

# Se pasa de minutos a horas y se cambia el tipo de dato de 'Playtime_forever' para que pese menos
df_final_genre['Hours']=(df_final_genre['Playtime_forever']/60).round().astype('int32')

# Se elimina la columna 'Playtime_forever' ya que no es necesaria
df_final_genre = df_final_genre.drop(columns='Playtime_forever')


df_final_genre.info()

# Se convierte a csv el DataFrame
df_final_genre.to_csv('end1.csv',index=False)

## --> UserForGenre
### ** DataFrames y Columnas  **


- A continuación se lista lo necesario para el segundo endpoint :

    2. def UserForGenre( genero : str ): Debe devolver el usuario que acumula más horas jugadas para el género dado y una lista de la acumulación de horas jugadas por año.

        
        - df_games.Genre = (str)
        - df_games.Id_game = (int)
        - df_items.Id_game = (int)
        - df_items.Playtime_forever = (str)
        - df_items.User_id = (str)
             
        A partir de los dataset obtener un nuevo conjunto de datos unicamente con los siguientes campos que necesitará la consulta:
        - user_id (str) = Id del usuario
        - playtime_hours (int) = cantidad de horas de juego totales
        - year (int) = año 

    ---> Con respecto a este endpoint dejo la inquietud dado a que lo se pide es muy ambiguo puesto que no se describe en ninguna parte el año en el que el usuario jugo cierta cantidad de horas de un determinado juego , por ende como la columna 'Playtime_forever' hace referencia al total de minutos jugados por un usuario, no los descrimina por año se sugiere cambien este endpoint por uno mas medible y real. para suplir el 'año' se tomara la fecha de lanzamiento de cada juego para asi tener un año de inicio, para coger los minutos totales y hacer el calculo de horas por año, se tiene como base que un jugador profesional entrena 6 a 8 horas diaras de lunes a sabado lo que significa al año juega un total de 1560 y 2496 horas, pero para fines mas practicos y no extender la tabla se define un tope de 2678 


### Playtime_forever por usuario

# Se cargan los DataFrame necesarios
df_games = pd.read_csv("./data/games.csv")
df_items = pd.read_csv("./data/items.csv")


# Secrea  una copia del df items y se asigna el otro nombre
time_user_game = df_items[['Id_game','User_id','Playtime_forever']].copy()
time_user_game.head(1)

# se eliminan vfilas con valores 0 en playtime
time_user_game = time_user_game[time_user_game["Playtime_forever"] != 0]
time_user_game = time_user_game.reset_index(drop=True)

time_user_game.head(1)

# se hace copia del df games con otro nombre
genres_id_year = df_games[['Id_game','Genres','Release_year']].copy()
genres_id_year.sample(10)

genres_id_year['Release_year'].unique()

# Se verifica las filas que tienen 'No_Data'
No_Data = (genres_id_year['Release_year'] == 'No_Data').sum()
No_Data

# Se eliminan las filas 'No_Data'
genres_id_year=genres_id_year[genres_id_year['Release_year'] != 'No_Data']

# cambio de tipo de la variable Releaseyear
genres_id_year['Release_year'] = genres_id_year['Release_year'].astype('int16')

# Convertir las cadenas tipo lista en listas reales y en filas separadas 
genres_id_year['Genres'] = genres_id_year['Genres'].apply(ast.literal_eval)
genres_id_year = genres_id_year.explode('Genres')


# se unen los dos df en uno solo 
df_userMerge =  genres_id_year.merge(time_user_game, on='Id_game', how='inner')
df_userMerge.head(2)

# se realiza una agrupacion sumando el tiempo de cada jugador por categoria
df_final_UserGenre = df_userMerge.groupby(['Genres', 'Release_year','Id_game','User_id'])['Playtime_forever'].sum().reset_index()

df_final_UserGenre.head(1)

# Se pasan de minutos a horas 
df_final_UserGenre['Hours']=(df_final_UserGenre['Playtime_forever']/60).round().astype('int32')

# Se eliminan filas en donde horas sea = 0
df_final_UserGenre = df_final_UserGenre[df_final_UserGenre['Hours'] != 0]

# Se explora el valor que tiene mas horas
df_final_UserGenre['Hours'].max()

# Se crea una copia de la columna hours para pruebas 
df_final_UserGenre['Original_Hours'] = df_final_UserGenre['Hours']

# Define una función personalizada para calcular la cantidad de horas por año ya que existen jugadores con muchas horas de juego
# se definio el tope como 2678 por año

def calcular_year2(row):
    if row['Hours'] > 2678: # si el valor es mayor a 278
        return row['Hours'] - 2678 # se asigana el valor de la resta entre el valor de la fila - 2678
    else:
        return 0  # Si no es asi coloca un 0 

# se ejecuta la funccion para la columna years que se creara
df_final_UserGenre['year2'] = df_final_UserGenre.apply(calcular_year2, axis=1) 

# Se establece el tope de horas por año a los usuarios que tengas mas de 2678 horas
df_final_UserGenre['Hours'] = df_final_UserGenre['Hours'].apply(lambda x: 2678 if x > 2678 else x)

# Define una función personalizada para calcular 'year3' haciendo el mismo proceso anterior

def calcular_year3(row):
    if row['year2'] > 2678:
        return row['year2'] - 2678
    else:
        return 0  # Si no cumple la condición, ponemos None

df_final_UserGenre['year3'] = df_final_UserGenre.apply(calcular_year3, axis=1)

df_final_UserGenre['year2'] = df_final_UserGenre['year2'].apply(lambda x: 2678 if x > 2678 else x)

# Define una función personalizada para calcular 'year4'

def calcular_year4(row):
    if row['year3'] > 2678:
        return row['year3'] - 2678
    else:
        return 0  # Si no cumple la condición, ponemos None

df_final_UserGenre['year4'] = df_final_UserGenre.apply(calcular_year4, axis=1)

df_final_UserGenre['year3'] = df_final_UserGenre['year3'].apply(lambda x: 2678 if x > 2678 else x)

# se cambia el nombre de la columna inial de hora para tener una similitud en las columnas
df_final_UserGenre = df_final_UserGenre.rename(columns={'Hours': 'year1'})

# Se cambian los tipos de variables de todas las columnas 
df_final_UserGenre['year1'] = df_final_UserGenre['year1'].astype('int16')
df_final_UserGenre['year2'] = df_final_UserGenre['year2'].astype('int16')
df_final_UserGenre['year3'] = df_final_UserGenre['year3'].astype('int16')
df_final_UserGenre['year4'] = df_final_UserGenre['year4'].astype('int16')

df_final_UserGenre.head(2)

# Se eliminan las columnas innecesarias 
df_final_UserGenre.drop(columns=['Id_game','Playtime_forever'], inplace=True)

# se ordenan las columnas a una manera mas legible
orden=['Genres','Release_year','User_id','Original_Hours','year1','year2','year3','year4',]
df_final_UserGenre = df_final_UserGenre[orden]

df_final_UserGenre

Debido a que es muy grande la data  se empieza a depurar para que no consuma muchos recursos en la API

# Se depuran las filas en donde los dos primeros años tengan 0
df_final_UserGenre = df_final_UserGenre[(df_final_UserGenre['year1'] != 0) & (df_final_UserGenre['year2'] != 0)]

df_final_UserGenre

# Se convierte a csv el DataFrame
df_final_UserGenre.to_csv('end2.csv',index=False)

## --> UsersRecommend y UsersNotRecommend
### ** DataFrames y Columnas  **


- A continuación se lista lo necesario para el tercer y cuarto endpoint :

    3.  def UsersRecommend( año : int ): evuelve el top 3 de juegos MÁS recomendados por usuarios para el año dado.
        
        - df_games.Id_game = (int)
        - df_games.Release_year = (int)
        - df_games.App_name = (str)
        - df_review.Id_game = (int)
        - df_review.Recommended = (bool)
        - df_review.Sentiment_analysis = (int)
             
        A partir de los dataset obtener un nuevo conjunto de datos unicamente con los siguientes campos que necesitará la consulta:
        - App_name (str) = nombre del juego
        - Release_Year (int) = Año de lanzamiento del juego
        - Recommend (int) = Recomentacion positva o negativa del juego 
        - Sentiment_analysis (int) = Analisis de sentimiento (separado en 3 columnas 'Positivo','Malo','Neutral')
        
        <br>
        <br>
    4.  def UsersNotRecommend( año : int ): Devuelve el top 3 de juegos MENOS recomendados por usuarios para el año dado.

        - df_games.Id_game = (int)
        - df_games.Release_year = (int)
        - df_games.App_name = (str)
        - df_review.Id_game = (int)
        - df_review.Recommended = (bool)
        - df_review.Sentiment_analysis = (int)
             
        A partir de los dataset obtener un nuevo conjunto de datos unicamente con los siguientes campos que necesitará la consulta:
        - App_name (str) = nombre del juego
        - Release_Year (int) = Año de lanzamiento del juego
        - Recommend (int) = Recomentacion positva o negativa del juego 
        - Sentiment_analysis (int) = Analisis de sentimiento (separado en 3 columnas 'Positivo','Malo','Neutral')
        

        


# Se cargan los DataFrame necesarios
df_games = pd.read_csv("./data/games.csv")
df_review = pd.read_csv("./data/review_sentiment.csv")

# Se crean los dataframes a utilizar para luego hacer el dataframe final para la consulta
df_g = df_games[['Id_game','App_name','Release_year']].copy()
df_r = df_review[['Id_game','Recommend','Sentiment_analysis',]].copy()

# Se realiza un inner join 
df_gr =  df_g.merge(df_r, on='Id_game', how='inner')
df_gr.head(10)

# Se crea una copia para el modelo de ML

df_gr_copia = df_gr[['Id_game','App_name','Recommend','Sentiment_analysis']]
df_gr_copia

# Se elimina la columna innecesario 'Id_game'
df_gr.drop(columns=['Id_game'], inplace=True)

# Se aplica un one-hot encoding a la columna 'Sentiment_analysis'
dummies_review = pd.get_dummies(df_gr['Sentiment_analysis'], prefix='Sentiment_analysis')

# Se concatena el df original con el df antteriormente creado 'dummies...'
df_gr = pd.concat([df_gr, dummies_review], axis=1)
df_gr


# Se elimina la columna 'Sentiment_analysis' y se renombrar unas columnas
df_gr = df_gr.drop('Sentiment_analysis', axis=1)
df_gr = df_gr.rename(columns={'Sentiment_analysis_0': 'Malo','Sentiment_analysis_1': 'Neutral','Sentiment_analysis_2': 'Positivo' })

# Se cambia el typo de dato a las columnas anteriormente nombradas para reducir el peso de df
df_gr['Malo'] = df_gr['Malo'].astype('int8')
df_gr['Neutral'] = df_gr['Neutral'].astype('int8')
df_gr['Positivo'] = df_gr['Positivo'].astype('int8')

df_gr.sample(20)

# Se convierte a csv el DataFrame
df_gr.to_csv('end3.csv',index=False)

## --> sentiment_analysis
### ** DataFrames y Columnas  **


- A continuación se lista lo necesario para el quinto endpoint :

    5. def sentiment_analysis( año : int ): Según el año de lanzamiento, se devuelve una lista con la cantidad de registros de reseñas categorizados con un análisis de sentimiento.
        
        - df_games.Release_year = (int)
        - df_games.Id_game = (int)
        - df_review.Id_game = (int)
        - df_review.Sentiment_analysis = (int)
             
        A partir de los dataset obtener un nuevo conjunto de datos unicamente con los siguientes campos que necesitará la consulta:

        - Release_Year (int) = año de lanzamiento del juego
        - Sentiment_analysis (int) = Analisis de sentimiento (separado en 3 columnas 'Positivo','Malo','Neutral')

# Se usa el DataFrame anteriormente creado y se usaran las columnas necesarias 
df_sent = df_gr[['Release_year','Malo','Neutral','Positivo']]

# Se agrupa el DataFrame por año y se suma los valores de cada columna por año 
df_sent_agru = df_sent = df_sent.groupby('Release_year', as_index=False).sum()
df_sent_agru.head(2)

# Se convierte a csv el DataFrame
df_sent_agru.to_csv('end5.csv',index=False)

## --> Modelo ML

Se creara un DataFrame con las columnas necesarias para el modelo de ML 

# Se crea el DF
ml = df_gr_copia
ml.head(3)

Con base a las columnas 'Recommend,Malo,Neutral y Positivo' se buscara crear una puntuacion para cada juego teniendo en cuenta la combinacion de estas columnas y se le asiganra una escala de la siguiente manera:

- **1** si en la columna Malo tiene (1) y Recommend tiene (1 o 0)
- **2** si en la columna Neutral tiene (1) y Recommend tiene (0)
- **3** si en la columna Neutral tiene (1) y Recommend (1)
- **4** si en la columna Positivo tiene (1) y Recommend (0)
- **5** si en la columna Positivo tiene (1) y Recommend (1)

Esto se hace para poder filtrar los mejores juegos

def escala(row):

    if row["Sentiment_analysis"] == 0 and not row["Recommend"]:
        return 1
    elif row["Sentiment_analysis"] == 0 and row["Recommend"]:
        return 1
    elif row["Sentiment_analysis"] == 1 and not row["Recommend"]:
        return 2
    elif row["Sentiment_analysis"] == 1 and row["Recommend"]:
        return 3
    elif row["Sentiment_analysis"] == 2 and not row["Recommend"]:
        return 4
    elif row["Sentiment_analysis"] == 2 and row["Recommend"]:
        return 5
    else:
        return None

ml['Score'] = ml.apply(escala,axis =1)
ml.sample(1)

# Se filtra el DF en donde solo se dejan los juegos con score 5
top_score = ml[ml['Score']>4]
top_score

# funcion para  agrupar por "Id_game" manteniendo el nombre del juego 
def first_app_name(series):
    return series.iloc[0]  # Obtenemos el primer valor de la serie

# se aplica la funcion al df grupo
top_score = top_score.groupby(["Id_game", "Score"])["App_name"].apply(first_app_name).reset_index()

top_score.sample(5)

# Se anexa la columna Genres del DF games 
top_score = top_score.merge(df_games[['Id_game','Genres']],on='Id_game', how='left')
top_score.sample(7)

# Se eliminan caracteres de la columna 'Genres'

def eliminar_caracteres(cadena):
    # Reemplaza los caracteres no deseados por una cadena vacía
    caracteres_a_eliminar = ["[", "]", "'", " ", '"']
    for caracter in caracteres_a_eliminar:
        cadena = cadena.replace(caracter, '')

    return cadena

top_score['Genres'] =top_score['Genres'].apply(eliminar_caracteres)

# se realiza una funcion para eliminar palabras repetidas 

def eliminar_palabras_repetidas(cadena):
    cadena = cadena.split(",")  # se divide la cadena por medio del seprador ","  quedando uan lista de palabras
    palabras_unicas = list(set(cadena))  # Se convierte la lista en conjunto y automaticamente elimina las palabras duplicadas
    texto_sin_repetidos = ','.join(palabras_unicas) # Se convierte el conjunto a cadena uniendo las palabras por "," y un espacio
    return texto_sin_repetidos

    # Se aplica 
top_score['Genres'] =top_score['Genres'].apply(eliminar_palabras_repetidas)

# Se pone en minuscula todo el contenido de la columna 'Genres'
top_score['Genres'] =top_score['Genres'].str.lower()

top_score.sample(5)

top_score

top_score.to_csv("datos_ML.csv")
